Use React instead and give up on Vue
vue's language server, or lsp-mode's integration of it (I'm somewhat
sure it's the former but not 100%) fails to anticipate when a
package.json is not at the root of the repository, so it (a) cannot find
@tanstack/vue-table and (b) does not see the type:module in
web/package.json and complains about imports.

This is more pain than just using React instead.