Give up on adding a loading state on react's side
it *really* doesn't want you to render something different. But what the
server sends is still necessarily different from the client displaying
the stored state. And the """solution""" is to make the client render
exactly like the server but immediately set the state afterwards in a
useEffect or something, which means you render twice on startup.

Or we can just, do it from Astro's side if we're not committing to
making this also work as a server application. Which we're not.