fix build
SQLite 3.50 added the unistr() functions, and they're also used in
dumps. epkg.sql now has unistr() calls and therefore requires sqlite
3.50+. The version in the CI image is 3.45.1.

We could fix this by installing an up-to-date sqlite with another action
or with brew, or we could just take advantage of bun's embedded sqlite,
which is already 3.50.

This is unfortunately far less efficient because .read actually makes
sqlite directly read from the file, whereas through a JS interface we
have to read the 21MB .sql plain text into memory first, and then pass
the entire thing as a string to db.run(). On my machine it went from <3s
to about a minute.