try to speed up epkg loading with a stream
maybe the bottleneck is reading the whole file into memory? I dunno.
This may not even work, since db.run creates its own statement.

Edit: yes, that was effective, at least in perf. On GitHub Actions it
took about 1 secons (or less) instead of 20 woth the previous approach.
There were no errors, and the resulting row count is the same, so this
is probably a success.